Refactoring: Use dedicated class FileInfo
We are now using a dedicated class FileInfo to store information about a
local or remote file. This makes getting such information easier (in
terms of amount of code to write to get a particular property of a
file). By having "custom" properties, we still allow storing arbitrary
backend specific information in such a record.