Downloads + Etag fix
In the WebDAVDownloadFileJob, we now request compression on the server
to be turned off. This is required because some server implementations
alter the etag - this means, we get different etags when listing remote
folder contents vs downloading one of the files.

With this fix, the DirectorySynchronizer now prefers the etags send
along the resource when downloading (vs the ones we obtained by the
earlier list files call, which might already be outdated).