Allow inclusion in other project via *.pri file
This change adds a *.pri file, which can be included by other projects
which use qmake to easily include this library. We also add a very
simple test which makes sure that a including project would at least
build (this is - of course - not a full validation of the functionality,
but should yield early warnings if something badly goes wrong.