Fix: Prevent apps from hanging during shutdown
This fixes apps using the SQLSyncStateDatabase class from potentially
hanging during shutdown. This was caused by the class holding an
instance of QSqlDatabase, which is heavily discouraged by Qt. Instead,
we now only save the connection name and use it whenever we need to get
an instance of the database connection.