Dropbox support - initial code drop
This commit adds a lot of boilerplate code for
adding Dropbox support to the library. All the
required job classes and tests are added. Some
of the jobs - the DropboxCreateFolderJob and
DropboxGetFileInfo job are already implemented.
The same is true for the DropboxCreateFolder
unit test.

Next up is implementing the remaining jobs
as well as their unit tests. Furthermore, the
DirectorySynchronizer needs some updates,
as Dropbox does not have "sync attributes"
(aka etags) on folders. However, their API
allows easily retrieving the list of changes
from the server in an elegant way. So we'll
extend the synchronizer (and job API) a bit
so we can utilize this approach in parallel
to the WebDAV scheme.