Add recusive listing + cursors
The ListFilesJob now has two additional properties:

The "recursive" property can be used to list a folder recursively (if
this is supported by the backend).
In addition, there is a new "cursor" property. After a completed
ListFilesJob, this property can be read and saved for later use. When
creating a new ListFilesJob, it's cursor can be set to the old value.
The intention is, that running the job will then only yield the changed
files and folders.

Both new properties are currently only used by the DropboxListFilesJob
implementation. The WebDAVListFilesJob could - in theory - also
implement recursive listing. However, this is done setting the "Depth"
HTTP header to "infinity", which not all servers support.

Both properties are required to implement the sync - as we cannot use
the same scheme as for WebDAV (as folders don't have a sync attribute).