Provide hint that the sync should be retried.
This change introduces a new boolean flag in the DirectorySynchronizer
class: retryWithFewerJobs. This flag is set to true after a failed sync
if the class detects that the sync could be retried with a reduced
number of parallel jobs.

This is a workaround for the problem reported via
https://github.com/mhoeher/opentodolist/issues/46. There, the server was
not able to handle all requests in parallel and hence prematurely closed
down connections. By re-running the sync with e.g. only a single job
running at a time, it worked just fine.