Allow WebDAV jobs against servers not reporting etags
This change allows running the WebDAV jobs against servers which don't
properly report back etags in the following cases:

- When downloading a file.
- When getting folder information about a single file or folder.

This is not yet enough to get the sync via the DirectorySynchronizer
working - this requires additional work. In particular, one of the
faulty servers (OpenDrive) pretty quickly responds with an HTTP 429 Too
Many Requests error. So probably, in addition to the etags, work will
need to be done in order to delay requests appropriately.