Fixes for Dropbox
This change adds some fixes needed to get the Dropbox sync up and
running in a downstream project. It basically fixes two minor issues:

- The path given to the filter is always sanitized (i.e. looks like an
absolute path).
- Instead of storing a bogus values as sync attribute, we store an empty
one in the sync state database. Otherwise, we will confuse the Dropbox
sync as it will try to pass the bogus string to the server (which
rightfully will reject it).